+++ pagegraph/__init__.py
@@ -1,3 +1,3 @@
 from packaging.version import Version
 
-VERSION = Version("0.6.3")
+VERSION = Version("0.6.4")