+++ pagegraph/serialize.py
@@ -1,5 +1,5 @@
 from dataclasses import dataclass, fields
-from typing import Any, cast
+from typing import Any, cast, Union
 
 from pagegraph.types import BlinkId, PageGraphId, Url, RequestId
 
@@ -9,17 +9,21 @@ class Report:
     pass
 
 
+JSONAble = Report | list[Report] | dict[str, Report] | str | int | float | bool
+
+
 @dataclass
 class FrameReport(Report):
-    nid: PageGraphId
+    id: PageGraphId
     url: Url | None
     blink_id: BlinkId
 
 
 @dataclass
 class DOMElementReport(Report):
-    nid: PageGraphId
+    id: PageGraphId
     tag: str
+    attrs: dict[str, JSONAble] | None = None
 
 
 @dataclass
@@ -36,13 +40,13 @@ class JSInvokeReport(Report):
 
 @dataclass
 class RequestReport(Report):
-    nid: PageGraphId
+    id: PageGraphId
     url: Url | None
 
 
 @dataclass
 class RequestCompleteReport(Report):
-    nid: PageGraphId
+    id: PageGraphId
     size: int
     hash: str
     headers: str
@@ -51,7 +55,7 @@ class RequestCompleteReport(Report):
 
 @dataclass
 class RequestErrorReport(Report):
-    nid: PageGraphId
+    id: PageGraphId
     headers: str | None
     status: str = "error"
 
@@ -65,6 +69,41 @@ class RequestChainReport(Report):
     result: RequestCompleteReport | RequestErrorReport | None
 
 
+@dataclass
+class ScriptReport(Report):
+    id: PageGraphId
+    type: str
+    hash: str
+    url: Url | None = None
+    source: str | None = None
+    executor: Union[DOMElementReport, "ScriptReport", None] = None
+
+
+@dataclass
+class ElementReport(Report):
+    id: PageGraphId
+    type: str
+    details: Union[dict[str, str], None]
+
+
+BriefNodeReport = ElementReport
+BriefEdgeReport = ElementReport
+
+
+@dataclass
+class NodeReport(ElementReport):
+    incoming_edges: list[Union["BriefEdgeReport", "EdgeReport", str]]
+    outgoing_edges: list[Union["BriefEdgeReport", "EdgeReport", str]]
+    kind: str = "node"
+
+
+@dataclass
+class EdgeReport(ElementReport):
+    incoming_node: Union["NodeReport", "BriefNodeReport", str, None]
+    outgoing_node: Union["NodeReport", "BriefNodeReport", str, None]
+    kind: str = "edge"
+
+
 class Reportable:
     def to_report(self) -> Report:
         raise NotImplementedError()
@@ -74,16 +113,15 @@ def report_field_name(field_name: str) -> str:
     return field_name.replace("_", " ")
 
 
-JSONAble = Report | list[Report] | dict[str, Report] | str | int | float | bool
-
-
 def to_jsonable(data: JSONAble) -> Any:
     if isinstance(data, list):
-        return [to_jsonable(x) for x in data]
+        return [to_jsonable(x) for x in data if x is not None]
 
     if isinstance(data, dict):
         jsonable_dict: dict[str, JSONAble] = {}
         for k, v in data.items():
+            if v is None:
+                continue
             report_key = report_field_name(k)
             jsonable_dict[report_key] = to_jsonable(v)
         return jsonable_dict
@@ -93,6 +131,8 @@ def to_jsonable(data: JSONAble) -> Any:
         for field in fields(data):
             field_name = field.name
             value = getattr(data, field_name)
+            if value is None:
+                continue
             report_name = report_field_name(field_name)
             jsonable_map[report_name] = to_jsonable(value)
         return jsonable_map